migrate table bootstraping to seperate file
